replaced ints with longs (ick), and made newgrid bigger in screen.cpp
git-svn-id: svn+ssh://svn.code.sf.net/p/snowstorm/code/openglad/trunk@1619 636a839e-d62e-0410-834d-a1d0aacdb804

+++ src/button.h
@@ -28,7 +28,7 @@
 //#include "int32.h"
 #include "input.h"
 
-#define MEM_TIME (long) 2000
+#define MEM_TIME (int) 2000
 class vbutton;
 
 // Definition of a button
@@ -50,18 +50,18 @@ class vbutton;
 //extern screen *myscreen;
 extern screen *myscreen;
 extern text *mytext;
-extern long *mymouse;
-extern unsigned long money[4];
+extern int *mymouse;
+extern unsigned int money[4];
 
 typedef struct
 {
 	char label[30];
 	unsigned char hotkey;
-	long x, y;
-	long sizex, sizey;
-	//long (*fun)(long arg1);
-	long myfun;
-	long arg1;     // argurment to function fun
+	int x, y;
+	int sizex, sizey;
+	//int (*fun)(int arg1);
+	int myfun;
+	int arg1;     // argurment to function fun
 }
 button;
 
@@ -69,34 +69,34 @@ class vbutton
 {
 	public:
 		vbutton();//this should only be used for pointers!!
-		vbutton(long xpos, long ypos, long wide, long high, long func(long),
-		        long pass, char *msg, unsigned char hot );
-		vbutton(long xpos, long ypos, long wide, long high, long func_code,
-		        long pass, char *msg, unsigned char hot );
-		vbutton(long xpos, long ypos, long wide, long high, long func_code,
-		        long pass, char *msg, char family, unsigned char hot );
+		vbutton(int xpos, int ypos, int wide, int high, int func(int),
+		        int pass, char *msg, unsigned char hot );
+		vbutton(int xpos, int ypos, int wide, int high, int func_code,
+		        int pass, char *msg, unsigned char hot );
+		vbutton(int xpos, int ypos, int wide, int high, int func_code,
+		        int pass, char *msg, char family, unsigned char hot );
 		~vbutton();
 		void set_graphic(char family);
-		long leftclick(); //is called when the button is left clicked
-		long leftclick(long whichone);
-		long rightclick(); //is called when the button is right clicked
-		long rightclick(long whichone);
-		long mouse_on(); //determins if mouse is on this button, returns 1 if true
+		int leftclick(); //is called when the button is left clicked
+		int leftclick(int whichone);
+		int rightclick(); //is called when the button is right clicked
+		int rightclick(int whichone);
+		int mouse_on(); //determins if mouse is on this button, returns 1 if true
 		void vdisplay();
-		void vdisplay(long status); // display depressed
-		long do_call(long whatfunc, long arg);
-		long do_call_right(long whatfunc, long arg);  // for right-button
+		void vdisplay(int status); // display depressed
+		int do_call(int whatfunc, int arg);
+		int do_call_right(int whatfunc, int arg);  // for right-button
 
-		long xloc; //the xposition in screen-coords
-		long yloc; //the yposition in screen-coords
+		int xloc; //the xposition in screen-coords
+		int yloc; //the yposition in screen-coords
 		char label[80]; //the label on the button
-		long width; // the buttons width in pixels
-		long height; // the buttons height in pixels
-		long xend; //xloc+width
-		long yend; //yloc+height
-		long (*fun)(long arg1); //the function this button calls when clicked, with one arg
-		long myfunc;
-		long arg; //the arg to be passed to the function when called
+		int width; // the buttons width in pixels
+		int height; // the buttons height in pixels
+		int xend; //xloc+width
+		int yend; //yloc+height
+		int (*fun)(int arg1); //the function this button calls when clicked, with one arg
+		int myfunc;
+		int arg; //the arg to be passed to the function when called
 		vbutton * next; //a pointer to the next button
 		vbutton * prev; //a pointer to the previous button
 		char had_focus; // did we recently have focus?
@@ -111,60 +111,60 @@ extern vbutton *allbuttons[MAX_BUTTONS];
 
 short has_mouse_focus(button thisbutton);
 void clearmenu(button *buttons, short numbuttons);
-long add_money(long howmuch);
+int add_money(int howmuch);
 
-long ventermenu(vbutton *vbuttons);
-long vexitmenu(vbutton *vbuttons);
+int ventermenu(vbutton *vbuttons);
+int vexitmenu(vbutton *vbuttons);
 
-vbutton * buttonmenu(button * buttons, long numbuttons);
-vbutton * buttonmenu(button * buttons, long numbuttons, long redraw);
+vbutton * buttonmenu(button * buttons, int numbuttons);
+vbutton * buttonmenu(button * buttons, int numbuttons, int redraw);
 
 // These are for picker ..
-long score_panel(screen *myscreen);
-long mainmenu(long arg1);
-long beginmenu(long arg1);
-long loadmenu(long arg1);
-long newmenu(long arg1);
-void quit(long arg1);
-long nullmenu(long arg1);
-long load1(long arg1);  // Begin a preset scenario ..
-long load2(long arg1);
-long load3(long arg1);
-long create_team_menu(long arg1); // Create / modify team members
-long create_detail_menu(guy *arg1); // detailed character information
-long create_view_menu(long arg1); // View team members
-long create_buy_menu(long arg1);  // Purchase new team members
-long create_edit_menu(long arg1); // Edit or sell team members
-long create_load_menu(long arg1); // Load a team
-long create_save_menu(long arg1); // Save a team
-long go_menu(long arg1); // run glad..
-long increase_stat(long arg1, long howmuch=1); // increase a guy's stats
-long decrease_stat(long arg1, long howmuch=1); // decrease a guy's stats
-unsigned long calculate_cost();
-unsigned long calculate_cost(guy * oldguy);
-long cycle_guy(long whichway);
-long cycle_team_guy(long whichway);
-long add_guy(long ignoreme);
-long edit_guy(long arg1); // transfer stats .. hardcoded
-long do_save(long arg1);  // dummy function for save_team_list
-long save_team_list(const char * filename); // save the team list
-long do_load(long arg1); // dummy function for load_team_list_one
-long load_team_list_one(const char * filename); // load a team list
-long delete_all(); // delete entire team
-long delete_first(); // delete first guy on team list
-long how_many(long whatfamily);   // how many guys of family X on the team?
+int score_panel(screen *myscreen);
+int mainmenu(int arg1);
+int beginmenu(int arg1);
+int loadmenu(int arg1);
+int newmenu(int arg1);
+void quit(int arg1);
+int nullmenu(int arg1);
+int load1(int arg1);  // Begin a preset scenario ..
+int load2(int arg1);
+int load3(int arg1);
+int create_team_menu(int arg1); // Create / modify team members
+int create_detail_menu(guy *arg1); // detailed character information
+int create_view_menu(int arg1); // View team members
+int create_buy_menu(int arg1);  // Purchase new team members
+int create_edit_menu(int arg1); // Edit or sell team members
+int create_load_menu(int arg1); // Load a team
+int create_save_menu(int arg1); // Save a team
+int go_menu(int arg1); // run glad..
+int increase_stat(int arg1, int howmuch=1); // increase a guy's stats
+int decrease_stat(int arg1, int howmuch=1); // decrease a guy's stats
+unsigned int calculate_cost();
+unsigned int calculate_cost(guy * oldguy);
+int cycle_guy(int whichway);
+int cycle_team_guy(int whichway);
+int add_guy(int ignoreme);
+int edit_guy(int arg1); // transfer stats .. hardcoded
+int do_save(int arg1);  // dummy function for save_team_list
+int save_team_list(const char * filename); // save the team list
+int do_load(int arg1); // dummy function for load_team_list_one
+int load_team_list_one(const char * filename); // load a team list
+int delete_all(); // delete entire team
+int delete_first(); // delete first guy on team list
+int how_many(int whatfamily);   // how many guys of family X on the team?
 void statscopy(guy *dest, guy *source); //copy stats from source => dest
-long set_player_mode(long howmany);
-long calculate_level(unsigned long temp_exp);
-unsigned long calculate_exp(long level);
+int set_player_mode(int howmany);
+int calculate_level(unsigned int temp_exp);
+unsigned int calculate_exp(int level);
 void clear_levels();
-long return_menu(long arg);
-long name_guy(long arg); // name the current guy
-long do_set_scen_level(long arg1);
-long set_difficulty();
-long change_teamnum(long arg);
-long change_hire_teamnum(long arg);
-long change_allied();
+int return_menu(int arg);
+int name_guy(int arg); // name the current guy
+int do_set_scen_level(int arg1);
+int set_difficulty();
+int change_teamnum(int arg);
+int change_hire_teamnum(int arg);
+int change_allied();
 
 // Function definitions ..
 #define BEGINMENU               1
